requirements.txt
fastapi
uvicorn
langchain
crewai
openai-whisper
vaderSentiment
coqui-tts
twilio
# sqlite3
python-dotenv
transformers
pandas 
chromadb 
sentence-transformers